finish up PartCR implementation
add some bells and whistles like auto-iteration based on timing info and modular design for heuristics

also, the "smart" RC splitter turned out to be dumb, so I removed most of the supporting code for that, but the splitter itself is still there if anyone wants to play with it later.

PartCR works pretty well now. On my simple test design problem, it prunes ~80% of low-energy conformations in a tiny fraction (~4%) of the time it would take to enumerate/minimize them.

oh, I also refactored the Stopwatch a bit so I can use fancy time formatting independently