optimize PartCR a bit, get rid of DEE step, and clean up a bit
The DEE step was an attempt to cut down on the time needed to resize the energy matrix, but the lazy evaluation seems to do a pretty good job of that all by itself. Also, I'm not entirely sure doing DEE on the bizzaro split-up conformation space is actually sound. Initial tests showed it was pruning too many things, so let's just get rid of it for now.

also, allow construction of RCs instance without pruning matrix