fix gpu memory management issues
the cpu side has a nice garbage collector, but the gpu side does not.
when cpu resources get garbage collected, I think the corresponding
gpu resources get released too. but the cpu-side has a lot more memory
than the gpu side, so the cpu-side might not get garbage collected
before the gpu-side runs out of memory. so we need to be a little more
aggressive about cleanup and explicitly call GpuEnergyFunction.cleanup()
when we're done with those instances.