upgrade SimpleCCDMinimizer to be faster
I replaced the line search part with Mark's super cool algorithm in
CCDMinimizer. My implementation is actually slightly faster than the
original one, but I'm having hard time reproducing the minimized
energies very precisely. Our parameterized molecule system appears to
be somewhat non-deterministic with respect to the DoF values, so it's
hard to reproduce energies exactly. It's pretty close though.

also, the eventual goal isn't to make CCD faster by just optimizing the
Java code. Eventually, I want to run CCD directly on the GPU for probably
another 2x speedup in minimization performance. once we can run the
minimizer directly on the GPU, we can skip a lot of the communication
overhead that happens when we try to run lots of energy functions inside
a loop. latency to the GPU is too high for my tastes, so I'm looking for
more way to hide it with various kinds of pipelining.

anyway, I needed a CCD implementation I understood REALLY WELL and works
at least fast as the one was have now before I can try to port it to the
GPU.