implement gpu kernel for quasi-Newton minimizer
this speeds up the energy function sampling portion of each
minimization step dramatically! =) roughly 2-3x speedup over sampling
points in serial

the sampling is used to make the local quadratic surface approximation
which is solved to find a direction for the update step

the parallel sampling QN-minimizer isn't terribly accurate yet, but it's
actually faster than the CCD implementation even though it takes more
steps to converge!

I finally found a way to overcome the kernel launch overhead problems
we get when using gpus in the CCD minimizer.