implement specialized CCD kernel
kernel just does CCD on forcefields and nothing else. only dihedral dofs
are supported for now, but it should be possible to add other dof types
in the future. I optimized the kernel a little bit too, but it's not
that much faster

also updated nvcc compiler settings for compatibility with newer nvidia
architectures