update simple energy matrix calculator
tried out the new Cuda CCD minimizer for the tiny energy matrix-sized
energy functions, but there's still too much overhead for the GPU to
go faster than the CPU. GPU speed is at about 80% now instead of 30%
though, so that's an improvement at least.

I'll leave the GPU energy matrix code in for now, but mark it as
deprecated though so no one actually tries to use it. Maybe it can be
made faster some day in the future.