load config files from resources, rather than files
let paths in config files be relative to the config file itself,
rather than the working directory. that way config files don't have
to worry about what directory they're in when specifying paths to
eg other config files or pdbs.

this lets the default config resources properly reference other resources

a bunch of stuff is still broken though (like paths to files in unit
tests), but this is very much a work in progress. I'll fix the paths
next time.