added SimpleGMECFinder and SimpleConfMinimizer
to run the find GMEC algorithm on the new ConfSpace system

eventually we'll get rid of these stupid "Simple" prefixes and rename
eg SimpleGMECFinder to just GMECFinder. But for now, the prefix serves
to keep the new and old implementations separate but also both
accessible.

also moved a lot of classes used by *GMECFinder to the new "gmec"
package, just to stay organized

and fixed a memory leak with OpenCL kernels
added finalizers to cleanup() classes too to try to detect and mitigate
other memory leaks in the future