don't make confEcalcs cleanup fragEcalcs
that's actually a bug, since fragment ecalcs can be resued in multiple places.

eventually, I should probably get rid of the old MinimizingConfEnergyCalculator and its
interface entirely, but that's a refactor for another day.

Since the new MinimizingConfEnergyCalculator composes with a frag Ecalc, there's not much
need to have different confEcalcs anymore. All the fancy minimization polymorphism is handled
by the fragEcalc now.