Pass the TaskContext as a parameter to the partition function creation.
Setting it once made a number of implicit assumptions.  The
`putTaskContext()` method assumes that there's a valid confDB instance
in the KStar class, and that that confDB instance doesn't change. Yet
every time one opens and closes the confDB the instance changes. So the
context should be set prior to the pfunc being calculated, not once for
all of the pfuncs.