Default to an empty array instead of a null array
There is a case in the ResidueForcefieldEnergyCuda constructor where a
residue has a confProblem so the entire conformation is considered
broken. Yet subsets of the conformation are still being used in making
the energy matrix. The subsets look at the resPairs array, which was
previously null if the conformation was broken.  To avoid the null
pointer exception, initialize the array as empty instead of null.