Make remaining kotlin unit tests pass
Restructure the code that called testConf() in TestConfSpaceCompiler to
use kotest's test() function directly, for some reason it didn't allow
us to create the tests in the testConf() function call.

Update the name of the fragments to include the chain in
TestCompiledAtomPairs.

TestMutation.kt was taking too long to run because of all of the calls
to shouldBe(), so instead use kotlin sets and compute set differences
for the same effect, which runs in respectable unit test time.